Ajout biere trois dame
