+++ src/main/java/ch/heigvd/res/chill/domain/AxelVallon/Boxer.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.AxelVallon;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class Duvel implements IProduct {
+public class Boxer implements IProduct {
 
-  public final static String NAME = "Duvel";
-  public final static BigDecimal PRICE = new BigDecimal(3.8);
+  public final static String NAME = "Boxer";
+  public final static BigDecimal PRICE = new BigDecimal(2.0);
 
   @Override
   public String getName() {
