+++ src/main/java/ch/heigvd/res/chill/domain/AxelVallon/Duvel.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.AxelVallon;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class LaCorne implements IProduct {
+public class Duvel implements IProduct {
 
-  public final static String NAME = "La Corne";
-  public final static BigDecimal PRICE = new BigDecimal(4.85);
+  public final static String NAME = "Duvel";
+  public final static BigDecimal PRICE = new BigDecimal(3.8);
 
   @Override
   public String getName() {
