+++ src/main/java/ch/heigvd/res/chill/domain/AxelVallon/TroisDame.java
@@ -1,0 +1,21 @@
+package ch.heigvd.res.chill.domain.AxelVallon;
+
+import ch.heigvd.res.chill.domain.IProduct;
+
+import java.math.BigDecimal;
+
+public class TroisDame implements IProduct {
+
+    public final static String NAME = "Trois dame";
+    public final static BigDecimal PRICE = new BigDecimal(2.0);
+
+    @Override
+    public String getName() {
+        return NAME;
+    }
+
+    @Override
+    public BigDecimal getPrice() {
+        return PRICE;
+    }
+}
