+++ src/test/java/ch/heigvd/res/chill/domain/AxelVallon/LaFrancheTest.java
@@ -1,4 +1,4 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.AxelVallon;
 
 import ch.heigvd.res.chill.domain.Bartender;
 import ch.heigvd.res.chill.protocol.OrderRequest;
@@ -9,22 +9,22 @@ import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class LaCorneTest {
+class LaFrancheTest {
 
   @Test
-  void thePriceAndNameForLaCorneShouldBeCorrect() {
-    LaCorne beer = new LaCorne();
-    assertEquals(beer.getName(), LaCorne.NAME);
-    assertEquals(beer.getPrice(), LaCorne.PRICE);
+  void thePriceAndNameForLaFrancheShouldBeCorrect() {
+    LaFranche beer = new LaFranche();
+    assertEquals(beer.getName(), LaFranche.NAME);
+    assertEquals(beer.getPrice(), LaFranche.PRICE);
   }
 
   @Test
   void aBartenderShouldAcceptAnOrderForLaFranche() {
     Bartender jane = new Bartender();
-    String productName = "ch.heigvd.res.chill.domain.wasadigi.LaCorne";
+    String productName = "ch.heigvd.res.chill.domain.wasadigi.LaFranche";
     OrderRequest request = new OrderRequest(3, productName);
     OrderResponse response = jane.order(request);
-    BigDecimal expectedTotalPrice = LaCorne.PRICE.multiply(new BigDecimal(3));
+    BigDecimal expectedTotalPrice = LaFranche.PRICE.multiply(new BigDecimal(3));
     assertEquals(expectedTotalPrice, response.getTotalPrice());
   }
